Add files via upload
I uploaded this application to the Shiny server along with the data supporting it.  The interactive dashboard can be accessed at: https://ofgtrial.shinyapps.io/RShinyInteractive/
This dashboard is a part of the Open for Good Initiative at the Centre for Impact at the University of California Los Angeles' Anderson School of Management that aims to make corporate ESG disclosures more accessible.